refactor (*): restructure repo
remove simple interview-style array algos
partition code from documentation
