+++ pyprep/data_structures/stack_queue/stack.py
