+++ pyprep/utility/node.py
